[Feature][Kurt] - Displayed available access roles on the root
